pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.bandit]
skips = ["B101", "B608"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'
extend-exclude = '''
(
\.git |
\.__pycache__ |
\.hg |
\.mypy_cache |
\.tox |
\.venv |
_build |
buck-out |
build |
dist |
data
)
'''

[tool.coverage.run]
branch = true

[tool.flake8]
max-line-length = 120
max-complexity = 10
per-file-ignores = "__init__.py:F401"
extend-ignore = ["E203","W503","E402"]
extend-immutable-calls = ["pathlib.Path", "Path"]

[tool.isort]
profile = "black"
filter_files = true
known_first_party = ["TestingProject"]
known_third_party = ["api", "dotenv", "fastapi", "locust", "numpy", "pandas", "requests"]

[tool.mypy]
plugins = "numpy.typing.mypy_plugin"
check_untyped_defs = true
show_error_codes = true
pretty = true
warn_unused_configs = true
allow_redefinition = true
install_types = false # Note: does not work with pre-commit
non_interactive = false # Note: if needs installing check it is in precommit for mypy
ignore_missing_imports = true

[tool.poetry]
name = "data-api"
version = "0.1.0"
description = ""
authors = ['Pedro Neves <pedrocaldeiraneves@gmail.com>']
packages = [{ include = "TestingProject" }]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9"
pandas = "^1.4.3"
numpy = "^1.23.1"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
mysql-connector-python="8.0.32"
python-dotenv = "0.21.1"

[tool.poetry.dev-dependencies]
flake8 = "5.0.4"
black = "22.6.0"
mypy = "0.971"
pydocstyle = "6.1.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
isort = "5.10.1"
pre-commit = "^2.20.0"
bandit = "1.7.4"
jupyter = "^1.0.0"
jupyterlab = "^3.4.3"
locust = "^2.12.2"
sphinx = "5.3.0"
sphinx-rtd-theme = "^1.0.0"
sphinxcontrib-napoleon = "^0.7"
sphinx_mdinclude = "^0.5.3"
requests = "^2.28.1"

[tool.pycodestyle]
max-line-length = 120

[tool.pydocstyle]
convention = "numpy"
add_ignore = ["D107","D203","D213","D413"]

[tool.pylint]
max-line-length = 120
max-args = 10
disable = ["C0103","R0903","W0108","W0107","W0511"]

[tool.pytest.ini_options]
filterwarnings = "ignore::DeprecationWarning"
addopts = "--doctest-modules"